refactor(bookmark): replace query service with direct repository access in BookmarkService
